cli/pyproject.toml
[project]
name = "tab-cli"
version = "0.4.5"
description = "Tab — a verb-shaped CLI agent for the Tab plugin ecosystem"
readme = "README.md"
license = { file = "../LICENSE" }
requires-python = ">=3.12,<3.13"
authors = [{ name = "4lt7ab" }]
dependencies = [
    "typer>=0.15.1",
    "pydantic-ai>=0.0.20",
    "grimoire-core",
    "pyyaml>=6.0",
    # Used by `tab ask --model ollama:<name>` to bypass pydantic-ai's
    # OpenAI-compat routing for local Ollama. Currently transitive via
    # grimoire; pinned here explicitly so the dependency stays visible
    # if grimoire's transitive set changes.
    "ollama>=0.4",
]

[project.scripts]
tab = "tab_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tab_cli"]

[tool.uv.sources]
grimoire-core = { git = "https://github.com/4lt7ab/grimoire.git", tag = "v0.0.1", subdirectory = "packages/grimoire-core" }

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
